requirements.txt
aiohttp==3.12.14
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beartype==0.16.4
bitrix24-rest==1.1.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.4
django-admin-interface==0.26.1
django-bootstrap-modal-forms==3.0.4
django-ckeditor==6.7.1
django-colorfield==0.10.1
django-filter==24.3
django-js-asset==2.1.0
django-modeladmin-reorder==0.3.1
django-resized==1.0.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
fast-bitrix24==1.5.16
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
frozenlist==1.4.1
icontract==2.6.6
idna==3.7
isort==5.12.0
mccabe==0.7.0
more-itertools==10.1.0
multidict==6.0.4
mysqlclient==2.2.0
numpy==2.0.2
pandas==2.2.3
pep8-naming==0.13.3
Pillow==10.0.0
pybitrix24==1.1.0
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2024.2
requests==2.32.4
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tqdm==4.66.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.5.0
yarl==1.9.4
phonenumbers~=9.0.12